+++ Source/searchOptions.py
@@ -287,4 +287,4 @@ def unitTest():
     testCase.setPBFunction(2)
     print(testCase.getPBFunction())
 
-unitTest()
+#unitTest()